requirements.txt
google-cloud==0.34.0
google-cloud-bigquery==3.3.5
google-cloud-storage==2.5.0
python-dotenv~=0.21.0